add typescript as a dependency, following the spec of tsutils. see commit description for WTF?!
See
* https://github.com/typescript-eslint/typescript-eslint/pull/1120
* https://github.com/jest-community/eslint-plugin-jest/issues/467
* https://github.com/typescript-eslint/typescript-eslint/issues/1287

eslint-plugin-jest
depends on typescript-eslint/experimental-utils, which
depends on typescript-eslint/typescript-estree, which
depends on tsutils, which
depends on typescript as a peer dependency

tsutils has been added as a dependency to
typescript-eslint/typescript-estree from 2.6.0 to 2.6.1 (yes, a patch bump!)
in https://github.com/typescript-eslint/typescript-eslint/pull/1120

I'm guessing the peer dependency requirement went unchecked because
typescript was already in their devDependencies for obvious reaons,
but also tsutils was used in other packages in the monorepo,
making it a trivial change.

The reactions to the change were

1. "ignore the warning" https://github.com/typescript-eslint/typescript-eslint/issues/1287#issuecomment-560022251
So tens, hundreds, thousands, more people should not pay attention to a warning.
In my case, we actually fail the build process if there are unmet peer dependencies.
Since many repos depend on eslint, this means that all those repos will start failing. So ignore I can't.

2. use yarn or npm 6.11+ which introduced peerDependenciesMeta
which marks tsutils' peerDependency on typescript as optional
https://github.com/jest-community/eslint-plugin-jest/issues/467#issuecomment-551516484
https://github.com/typescript-eslint/typescript-eslint/pull/990
Except when running npm 6.12.1 and 6.14.11, I still get that error...

So that's why.